chore: reuse existing test file and fail test

+++ src/logic/prepare-e2e.ts
@@ -37,8 +37,13 @@ export async function prepareE2e(
   }
 
   const testPath = `./test/${E2E_DIR}/${scope}/${profile}/${provider}${EXTENSIONS.e2e}`;
-  //Create test file
-  await writeFile(testPath, e2eTestTemplate(scope, profile, provider));
+  //Reuse existing test file - so we can E2E test both staging and production.
+  if (await exists(testPath)) {
+    options?.warnCb?.(`E2E test file "${testPath}" already exists - reusing it`)
+  } else {
+    options?.logCb?.(`Creating test file "${testPath}"`)
+    await writeFile(testPath, e2eTestTemplate(scope, profile, provider));
+  }
 
   //Resolve api url
   const envUrl = process.env[SF_API_URL_VARIABLE];
